add code

+++ scripts/preprocess/unix.sh
@@ -177,6 +177,8 @@ if [[ -n ${uninstalled_pip_pkgs+x} ]]; then
     done
   done
 fi
-if test -f $SRC_DIR/.gitignore; then
-  rm $SRC_DIR/.gitignore
+if [[ -n ${SRC_DIR+x} ]]; then
+  if test -f $SRC_DIR/.gitignore; then
+    rm $SRC_DIR/.gitignore
+  fi
 fi
